Se arreglaron errores de sintaxis y compatibilidad con data-test.sql

+++ src/main/java/com/example/unirideapi/service/impl/RutaServiceImpl.java
@@ -248,7 +248,7 @@ public class RutaServiceImpl implements RutaService {
     @Transactional
     @Override
     public RutaResponseDTO updateEstadoRuta(Integer idRuta, EstadoRuta nuevoEstado) {
-        var ruta = rutaRepository.findById(idRuta)
+        var ruta = rutaRepository.findById((long)idRuta)
                 .orElseThrow(() -> new ResourceNotFoundException("Ruta no encontrada"));
 
         //Regla 1: Solo rutas PROGRAMADAS pueden cambiar a CONFIRMADO o CANCELADO
